chore: rename clipboard utils
