+++ packages/blocks/src/__internal__/clipboard/index.ts
@@ -1,4 +1,4 @@
 export * from './edgeless-clipboard.js';
 export * from './page-clipboard.js';
 export * from './type.js';
-export * from './util/utils.js';
+export * from './utils/commons.js';
