--- packages/blocks/src/__internal__/clipboard/util/index.ts
@@ -1,3 +1,0 @@
-export * from './markdown-utils.js';
-export * from './pure-util.js';
-export * from './utils.js';
