+++ packages/blocks/src/__internal__/clipboard/utils/pure.ts
