+++ packages/blocks/src/__internal__/content-parser/index.ts
@@ -4,7 +4,7 @@ import { Slot } from '@blocksuite/store';
 import { marked } from 'marked';
 
 import type { PageBlockModel } from '../../models.js';
-import { getFileFromClipboard } from '../clipboard/util/pure-util.js';
+import { getFileFromClipboard } from '../clipboard/utils/pure.js';
 import type { SerializedBlock } from '../utils/index.js';
 import { FileExporter } from './file-exporter/file-exporter.js';
 import { HtmlParser } from './parse-html.js';
